[velar]merge GKI android14-6.1-2025-05_r2 for LTS upgrade
PROPAGATED_from (CR)

Release date: 2025-05-22
TAG: android14-6.1-2025-05_r2
SPL: 2025-06-05
Deprecation date: 2026-06-01

kernel Artifacts: https://ci.android.com/builds/submitted/13536476/kernel_aarch64/latest
source tag: https://android.googlesource.com/kernel/common/+/refs/tags/android14-6.1-2025-05_r2
release URL: https://source.android.com/docs/core/architecture/kernel/gki-android14-6_1-release-builds

Conflicts:
	drivers/android/vendor_hooks.c
	drivers/bus/mhi/host/pm.c
	drivers/clk/qcom/clk-alpha-pll.c
	drivers/clk/qcom/clk-alpha-pll.h
	drivers/clk/qcom/clk-rcg2.c
	drivers/clk/qcom/clk-rpmh.c
	drivers/clk/qcom/dispcc-sm8250.c
	drivers/clk/qcom/gcc-sc8180x.c
	drivers/edac/qcom_edac.c
	drivers/iommu/arm/arm-smmu/arm-smmu.c
	drivers/md/dm-bow.c
	drivers/mmc/host/sdhci-msm.c
	drivers/remoteproc/qcom_q6v5_mss.c
	drivers/remoteproc/qcom_q6v5_pas.c
	drivers/rpmsg/qcom_glink_native.c
	drivers/slimbus/qcom-ngd-ctrl.c
	drivers/soc/qcom/cmd-db.c
	drivers/soc/qcom/rpmh-rsc.c
	drivers/soc/qcom/rpmh.c
	drivers/spi/spidev.c
	drivers/ufs/core/ufshcd.c
	drivers/ufs/host/ufs-qcom.c
	fs/f2fs/file.c
	fs/f2fs/inode.c
	fs/f2fs/segment.c
	include/dt-bindings/clock/qcom,gcc-sc8180x.h
	include/linux/soc/qcom/geni-se.h
	include/soc/qcom/cmd-db.h
	include/trace/hooks/mm.h
	net/qrtr/af_qrtr.c
	net/qrtr/ns.c

Reviewed-on: https://gerrit.mot.com/3457382
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Submit-Approved: Jira Key
(cherry picked from commit 4e67453fd65d69e30fc78fcbf145b63356de714a)
Reviewed-on: https://gerrit.mot.com/3496494
SLTApproved: Anandappan Chakravarthy <pjwt34@motorola.com>
Submit-Approved: Anandappan Chakravarthy <pjwt34@motorola.com>
Reviewed-on: https://gerrit.mot.com/3524327
Submit-Approved: Andrea dos Santos <andreads@motorola.com>

+++ sound/soc/intel/avs/apl.c
@@ -112,7 +112,7 @@ static int apl_coredump(struct avs_dev *adev, union avs_notify_msg *msg)
 	struct apl_log_buffer_layout layout;
 	void __iomem *addr, *buf;
 	size_t dump_size;
-	u16 offset = 0;
+	u32 offset = 0;
 	u8 *dump, *pos;
 
 	dump_size = AVS_FW_REGS_SIZE + msg->ext.coredump.stack_dump_size;
